refactor: extract duplicated safeParse and redisStatus helpers
